feat(CheatGiftService): extract cheat gift logic
refactor(GameService): delegate cheat gift methods to CheatGiftService
